Control.Distributed.Process.Platform

Description

Cloud Haskell Platform

It is important not to be too general when catching exceptions in handler code, because asynchonous exceptions provide cloud haskell with its process termination mechanism. Two exceptions in particular, signal the instigators intention to stop a process immediately, these are raised in response to the kill and exit primitives provided by the base distributed-process package.

Exported Types

class Addressable a where

Provides a unified API for addressing processes

Methods

sendTo :: Serializable m => a -> m -> Process ()

Send a message to the target asynchronously

resolve :: a -> Process (Maybe ProcessId)

Resolve the reference to a process id, or Nothing if resolution fails

data Recipient

Constructors

Pid ProcessId

Registered String

RemoteRegistered String NodeId

data TerminateReason

Provides a reason for process termination.

Constructors

TerminateNormal
indicates normal exit

TerminateShutdown
normal response to a Shutdown

TerminateOther !String
abnormal (error) shutdown

type Tag = Int

Tags provide uniqueness for messages, so that they can be matched with their response.

type TagPool = MVar Tag

Generates unique Tag for messages and response pairs. Each process that depends, directly or indirectly, on the call mechanisms in Control.Distributed.Process.Global.Call should have at most one TagPool on which to draw unique message tags.

Utilities and Extended Primitives

spawnLinkLocal :: Process () -> Process ProcessId

Node local version of spawnLink. Note that this is just the sequential composition of spawn and link. (The Unified semantics that underlies Cloud Haskell does not even support a synchronous link operation)

spawnMonitorLocal :: Process () -> Process (ProcessId, MonitorRef)

Like spawnLinkLocal, but monitor the spawned process

linkOnFailure :: ProcessId -> Process ()

CH's link primitive, unlike Erlang's, will trigger when the target process dies for any reason. This function has semantics like Erlang's: it will trigger ProcessLinkException only when the target dies abnormally.

times :: Int -> Process () -> Process ()

Apply the supplied expression n times

matchCond :: Serializable a => (a -> Maybe (Process b)) -> Match b

An alternative to matchIf that allows both predicate and action to be expressed in one parameter.

Call/Tagging support

newTagPool :: Process TagPool

Create a new per-process source of unique message identifiers.

getTag :: TagPool -> Process Tag

Extract a new identifier from a TagPool.

Registration and Process Lookup

whereisOrStart :: String -> Process () -> Process ProcessId

Returns the pid of the process that has been registered under the given name. This refers to a local, per-node registration, not global registration. If that name is unregistered, a process is started. This is a handy way to start per-node named servers.

whereisOrStartRemote :: NodeId -> String -> Closure (Process ()) -> Process (Maybe ProcessId)

A remote equivalent of whereisOrStart. It deals with the node registry on the given node, and the process, if it needs to be started, will run on that node. If the node is inaccessible, Nothing will be returned.

__remoteTable :: RemoteTable -> RemoteTable