Control.Distributed.Process.Platform.Async

Description

The async APIs provided by distributed-process-platform provide means for spawning asynchronous operations, waiting for their results, cancelling them and various other utilities. The two primary implementation are AsyncChan which provides a handle which is scoped to the calling process, and AsyncSTM, whose async mechanism can be used by (i.e., shared across) multiple local processes.

Both abstractions can run asynchronous operations on remote nodes. The STM based implementation provides a slightly richer API. The API defined in this module only supports a subset of operations on async handles, and (specifically) does not support mixing handles initialised via different implementations.

There is an implicit contract for async workers; Workers must exit normally (i.e., should not call the exit, die or terminate Cloud Haskell primitives), otherwise the AsyncResult will end up being AsyncFailed DiedException instead of containing the result.

See Control.Distributed.Process.Platform.Async.AsyncSTM, Control.Distributed.Process.Platform.Async.AsyncChan.

See Control.Distributed.Platform.Task for a high level layer built on these capabilities.

Exported Types

data Async a

An opaque handle that refers to an asynchronous operation.

type AsyncRef = ProcessId

A reference to an asynchronous action

data AsyncTask a

A task to be performed asynchronously.

Constructors

AsyncTask

Fields

asyncTask :: Process a

the task to be performed

AsyncRemoteTask

Fields

asyncTaskDict :: Static (SerializableDict a)

the serializable dict required to spawn a remote process

asyncTaskNode :: NodeId

the node on which to spawn the asynchronous task

asyncTaskProc :: Closure (Process a)

the task to be performed, wrapped in a closure environment

data AsyncResult a

Represents the result of an asynchronous action, which can be in one of several states at any given time.

Constructors

AsyncDone a
a completed action and its result

AsyncFailed DiedReason
a failed action and the failure reason

AsyncLinkFailed DiedReason
a link failure and the reason

AsyncCancelled
a cancelled action

AsyncPending
a pending action (that is still running)

Spawning asynchronous operations

async :: Serializable a => Process a -> Process (Async a)

Spawn an AsyncTask and return the Async handle to it. See asyncSTM.

asyncLinked :: Serializable a => Process a -> Process (Async a)

Spawn an AsyncTask (linked to the calling process) and return the Async handle to it. See asyncSTM.

asyncSTM :: Serializable a => AsyncTask a -> Process (Async a)

Spawn an AsyncTask and return the Async handle to it. Uses the STM implementation, whose handles can be read by other processes, though they're not Serializable.

See AsyncSTM.

asyncLinkedSTM :: Serializable a => AsyncTask a -> Process (Async a)

Spawn an AsyncTask (linked to the calling process) and return the Async handle to it. Uses the STM based implementation, whose handles can be read by other processes, though they're not Serializable.

See AsyncSTM.

asyncChan :: Serializable a => AsyncTask a -> Process (Async a)

Spawn an AsyncTask and return the Async handle to it. Uses a channel based implementation, whose handles can only be read once, and only by the calling process.

See AsyncChan.

asyncLinkedChan :: Serializable a => AsyncTask a -> Process (Async a)

Linked version of asyncChan.

See AsyncChan.

task :: Process a -> AsyncTask a

Wraps a regular Process a as an AsyncTask.

remoteTask :: Static (SerializableDict a) -> NodeId -> Closure (Process a) -> AsyncTask a

Wraps the components required and builds a remote AsyncTask.

Cancelling asynchronous operations

cancel :: Async a -> Process ()

Cancel an asynchronous operation. Cancellation is asynchronous in nature. To wait for cancellation to complete, use cancelWait instead. The notes about the asynchronous nature of cancelWait apply here also.

See Process

cancelWait :: Serializable a => Async a -> Process (AsyncResult a)

Cancel an asynchronous operation and wait for the cancellation to complete. Because of the asynchronous nature of message passing, the instruction to cancel will race with the asynchronous worker, so it is entirely possible that the AsyncResult returned will not necessarily be AsyncCancelled. For example, the worker may complete its task after this function is called, but before the cancellation instruction is acted upon.

If you wish to stop an asychronous operation immediately (with caveats) then consider using cancelWith or cancelKill instead.

cancelWith :: Serializable b => b -> Async a -> Process ()

Cancel an asynchronous operation immediately. This operation is performed by sending an exit signal to the asynchronous worker, which leads to the following semantics:

If the worker already completed, this function has no effect.

The worker might complete after this call, but before the signal arrives.

The worker might ignore the exit signal using catchExit.

In case of (3), this function has no effect. You should use cancel if you need to guarantee that the asynchronous task is unable to ignore the cancellation instruction.

You should also consider that when sending exit signals to a process, the definition of immediately is somewhat vague and a scheduler might take time to handle the request, which can lead to situations similar to (1) as listed above, if the scheduler to which the calling process' thread is bound decides to GC whilst another scheduler on which the worker is running is able to continue.

See exit

cancelKill :: String -> Async a -> Process ()

Like cancelWith but sends a kill instruction instead of an exit signal.

See kill

Querying for results

poll :: Serializable a => Async a -> Process (AsyncResult a)

Check whether an Async handle has completed yet. The status of the action is encoded in the returned AsyncResult. If the action has not completed, the result will be AsyncPending, or one of the other constructors otherwise. This function does not block waiting for the result. Use wait or waitTimeout if you need blocking/waiting semantics.

check :: Serializable a => Async a -> Process (Maybe (AsyncResult a))

Like poll but returns Nothing if (poll hAsync) == AsyncPending. See poll.

wait :: Async a -> Process (AsyncResult a)

Wait for an asynchronous action to complete, and return its value. The result (which can include failure and/or cancellation) is encoded by the AsyncResult type.

Waiting with timeouts

waitTimeout :: Serializable a => TimeInterval -> Async a -> Process (Maybe (AsyncResult a))

Wait for an asynchronous operation to complete or timeout. Returns Nothing if the AsyncResult does not change from AsyncPending within the specified delay, otherwise Just asyncResult is returned. If you want to wait/block on the AsyncResult without the indirection of Maybe then consider using wait or waitCheckTimeout instead.

waitCancelTimeout :: Serializable a => TimeInterval -> Async a -> Process (AsyncResult a)

Wait for an asynchronous operation to complete or timeout. If it times out, then cancelWait the async handle instead.

waitCheckTimeout :: Serializable a => TimeInterval -> Async a -> Process (AsyncResult a)

Wait for an asynchronous operation to complete or timeout. This variant returns the AsyncResult itself, which will be AsyncPending if the result has not been made available, otherwise one of the other constructors.