Control.Distributed.Process.Platform.Async.AsyncChan

Description

This module provides a set of operations for spawning Process operations and waiting for their results. It is a thin layer over the basic concurrency operations provided by Control.Distributed.Process. The main feature it provides is a pre-canned set of APIs for waiting on the result of one or more asynchronously running (and potentially distributed) processes.

The async handles returned by this module cannot be used by processes other than the caller of async, and are not Serializable. Specifically, calls that block until an async worker completes (i.e., all variants of wait) will never return if called from a different process.

 h <- newEmptyMVar
 outer <- spawnLocal $ async runMyAsyncTask >>= liftIO $ putMVar h
 hAsync <- liftIO $ takeMVar h
 say "this expression will never return, because hAsync belongs to 'outer'"
 wait hAsync

As with Async, workers can be started on a local or remote node.

See Control.Distributed.Platform.Async.

Exported types

type AsyncRef = ProcessId

A reference to an asynchronous action

data AsyncTask a

A task to be performed asynchronously.

Constructors

AsyncTask

Fields

asyncTask :: Process a

the task to be performed

AsyncRemoteTask

Fields

asyncTaskDict :: Static (SerializableDict a)

the serializable dict required to spawn a remote process

asyncTaskNode :: NodeId

the node on which to spawn the asynchronous task

asyncTaskProc :: Closure (Process a)

the task to be performed, wrapped in a closure environment

data AsyncChan a

A handle for an asynchronous action spawned by async. Asynchronous actions are run in a separate process, and operations are provided for waiting for asynchronous actions to complete and obtaining their results (see e.g. wait).

Handles of this type cannot cross remote boundaries. Furthermore, handles of this type must not be passed to functions in this module by processes other than the caller of async - that is, this module provides asynchronous actions whose results are accessible *only* by the initiating process. This limitation is imposed becuase of the use of typed channels, for which the ReceivePort component is effectively thread local.

See async

data AsyncResult a

Represents the result of an asynchronous action, which can be in one of several states at any given time.

Constructors

AsyncDone a
a completed action and its result

AsyncFailed DiedReason
a failed action and the failure reason

AsyncLinkFailed DiedReason
a link failure and the reason

AsyncCancelled
a cancelled action

AsyncPending
a pending action (that is still running)

data Async a

An opaque handle that refers to an asynchronous operation.

Spawning asynchronous operations

async :: Serializable a => AsyncTask a -> Process (AsyncChan a)

Spawns an asynchronous action in a new process. We ensure that if the caller's process exits, that the worker is killed. Because an AsyncChan can only be used by the initial caller's process, if that process dies then the result (if any) is discarded. If a process other than the initial caller attempts to obtain the result of an asynchronous action, the behaviour is undefined. It is highly likely that such a process will block indefinitely, quite possible that such behaviour could lead to deadlock and almost certain that resource starvation will occur. Do Not share the handles returned by this function across multiple processes.

If you need to spawn an asynchronous operation whose handle can be shared by multiple processes then use the AsyncSTM module instead.

There is currently a contract for async workers, that they should exit normally (i.e., they should not call the exit or kill with their own ProcessId nor use the terminate primitive to cease functining), otherwise the AsyncResult will end up being AsyncFailed DiedException instead of containing the desired result.

asyncLinked :: Serializable a => AsyncTask a -> Process (AsyncChan a)

For *AsyncChan*, async already ensures an AsyncChan is never left running unintentionally. This function is provided for compatibility with other async implementations that may offer different semantics for async with regards linking.

asyncLinked = async

newAsync :: Serializable a => (AsyncTask a -> Process (AsyncChan a)) -> AsyncTask a -> Process (Async a)

Create a new AsyncChane and wrap it in an Async record.

Used by Control.Distributed.Process.Platform.Async.

Cancelling asynchronous operations

cancel :: AsyncChan a -> Process ()

Cancel an asynchronous operation. Cancellation is asynchronous in nature.

See Control.Distributed.Process.Platform.Async.

cancelWait :: Serializable a => AsyncChan a -> Process (AsyncResult a)

Cancel an asynchronous operation and wait for the cancellation to complete.

See Control.Distributed.Process.Platform.Async.

cancelWith :: Serializable b => b -> AsyncChan a -> Process ()

Cancel an asynchronous operation immediately.

See Control.Distributed.Process.Platform.Async.

cancelKill :: String -> AsyncChan a -> Process ()

Like cancelWith but sends a kill instruction instead of an exit.

See Control.Distributed.Process.Platform.Async.

Querying for results

poll :: Serializable a => AsyncChan a -> Process (AsyncResult a)

Check whether an AsyncChan has completed yet.

See Control.Distributed.Process.Platform.Async.

check :: Serializable a => AsyncChan a -> Process (Maybe (AsyncResult a))

Like poll but returns Nothing if (poll hAsync) == AsyncPending.

See Control.Distributed.Process.Platform.Async.

wait :: Serializable a => AsyncChan a -> Process (AsyncResult a)

Wait for an asynchronous action to complete, and return its value. The outcome of the action is encoded as an AsyncResult.

See Control.Distributed.Process.Platform.Async.

waitAny :: Serializable a => [AsyncChan a] -> Process (AsyncResult a)

Wait for any of the supplied AsyncChanss to complete. If multiple Asyncs complete, then the value returned corresponds to the first completed Async in the list. Only unread Asyncs are of value here, because AsyncChan does not hold on to its result after it has been read!

This function is analagous to the mergePortsBiased primitive.

See Control.Distibuted.Process.mergePortsBiased.

waitAnyCancel :: Serializable a => [AsyncChan a] -> Process (AsyncResult a)

Like waitAny, but also cancels the other asynchronous operations as soon as one has completed.

Waiting with timeouts

waitAnyTimeout :: Serializable a => TimeInterval -> [AsyncChan a] -> Process (Maybe (AsyncResult a))

Like waitAny but times out after the specified delay.

waitTimeout :: Serializable a => TimeInterval -> AsyncChan a -> Process (Maybe (AsyncResult a))

Wait for an asynchronous operation to complete or timeout.

See Control.Distributed.Process.Platform.Async.

waitCancelTimeout :: Serializable a => TimeInterval -> AsyncChan a -> Process (AsyncResult a)

Wait for an asynchronous operation to complete or timeout. If it times out, then cancelWait the async handle instead.

waitCheckTimeout :: Serializable a => TimeInterval -> AsyncChan a -> Process (AsyncResult a)

Wait for an asynchronous operation to complete or timeout.

See Control.Distributed.Process.Platform.Async.