Control.Distributed.Process.Platform.Async.AsyncSTM

Description

This module provides a set of operations for spawning Process operations and waiting for their results. It is a thin layer over the basic concurrency operations provided by Control.Distributed.Process.

The difference between AsyncSTM and AsyncChan is that handles of the former (i.e., returned by this module) can be used by processes other than the caller of async, but are not Serializable.

As with AsyncChan, workers can be started on a local or remote node.

Portions of this file are derived from the Control.Concurrent.Async module, written by Simon Marlow.

Exported types

type AsyncRef = ProcessId

A reference to an asynchronous action

data AsyncTask a

A task to be performed asynchronously.

Constructors

AsyncTask

Fields

asyncTask :: Process a

the task to be performed

AsyncRemoteTask

Fields

asyncTaskDict :: Static (SerializableDict a)

the serializable dict required to spawn a remote process

asyncTaskNode :: NodeId

the node on which to spawn the asynchronous task

asyncTaskProc :: Closure (Process a)

the task to be performed, wrapped in a closure environment

data AsyncSTM a

An handle for an asynchronous action spawned by async. Asynchronous operations are run in a separate process, and operations are provided for waiting for asynchronous actions to complete and obtaining their results (see e.g. wait).

Handles of this type cannot cross remote boundaries, nor are they Serializable.

data AsyncResult a

Represents the result of an asynchronous action, which can be in one of several states at any given time.

Constructors

AsyncDone a
a completed action and its result

AsyncFailed DiedReason
a failed action and the failure reason

AsyncLinkFailed DiedReason
a link failure and the reason

AsyncCancelled
a cancelled action

AsyncPending
a pending action (that is still running)

data Async a

An opaque handle that refers to an asynchronous operation.

Spawning asynchronous operations

async :: Serializable a => AsyncTask a -> Process (AsyncSTM a)

Spawns an asynchronous action in a new process.

asyncLinked :: Serializable a => AsyncTask a -> Process (AsyncSTM a)

This is a useful variant of async that ensures an AsyncChan is never left running unintentionally. We ensure that if the caller's process exits, that the worker is killed. Because an AsyncChan can only be used by the initial caller's process, if that process dies then the result (if any) is discarded.

newAsync :: Serializable a => (AsyncTask a -> Process (AsyncSTM a)) -> AsyncTask a -> Process (Async a)

Create a new AsyncSTM and wrap it in an Async record.

Used by Async.

Cancelling asynchronous operations

cancel :: AsyncSTM a -> Process ()

Cancel an asynchronous operation.

See Control.Distributed.Process.Platform.Async.

cancelWait :: Serializable a => AsyncSTM a -> Process (AsyncResult a)

Cancel an asynchronous operation and wait for the cancellation to complete.

See Control.Distributed.Process.Platform.Async.

cancelWith :: Serializable b => b -> AsyncSTM a -> Process ()

Cancel an asynchronous operation immediately.

See Control.Distributed.Process.Platform.Async.

cancelKill :: String -> AsyncSTM a -> Process ()

Like cancelWith but sends a kill instruction instead of an exit.

See Async.

Querying for results

poll :: Serializable a => AsyncSTM a -> Process (AsyncResult a)

Check whether an AsyncSTM has completed yet.

See Control.Distributed.Process.Platform.Async.

check :: Serializable a => AsyncSTM a -> Process (Maybe (AsyncResult a))

Like poll but returns Nothing if (poll hAsync) == AsyncPending.

See Control.Distributed.Process.Platform.Async.

wait :: AsyncSTM a -> Process (AsyncResult a)

Wait for an asynchronous action to complete, and return its value. The result (which can include failure and/or cancellation) is encoded by the AsyncResult type.

wait = liftIO . atomically . waitSTM

See Control.Distributed.Process.Platform.Async.

waitAny :: Serializable a => [AsyncSTM a] -> Process (AsyncSTM a, AsyncResult a)

Wait for any of the supplied AsyncSTMs to complete. If multiple Asyncs complete, then the value returned corresponds to the first completed Async in the list.

NB: Unlike AsyncChan, AsyncSTM does not discard its AsyncResult once read, therefore the semantics of this function are different to the former. Specifically, if asyncs = [a1, a2, a3] and (AsyncDone _) = a1 then the remaining a2, a3 will never be returned by waitAny.

Waiting with timeouts

waitAnyTimeout :: Serializable a => TimeInterval -> [AsyncSTM a] -> Process (Maybe (AsyncResult a))

Like waitAny but times out after the specified delay.

waitTimeout :: Serializable a => TimeInterval -> AsyncSTM a -> Process (Maybe (AsyncResult a))

Wait for an asynchronous operation to complete or timeout.

See Control.Distributed.Process.Platform.Async.

waitCheckTimeout :: Serializable a => TimeInterval -> AsyncSTM a -> Process (AsyncResult a)

Wait for an asynchronous operation to complete or timeout.

See Control.Distributed.Process.Platform.Async.

STM versions

pollSTM :: AsyncSTM a -> STM (Maybe (AsyncResult a))

A version of poll that can be used inside an STM transaction.

waitTimeoutSTM :: Serializable a => TimeInterval -> AsyncSTM a -> Process (Maybe (AsyncResult a))

As waitTimeout but uses STM directly, which might be more efficient.

waitAnyCancel :: Serializable a => [AsyncSTM a] -> Process (AsyncSTM a, AsyncResult a)

Like waitAny, but also cancels the other asynchronous operations as soon as one has completed.

waitEither :: AsyncSTM a -> AsyncSTM b -> Process (Either (AsyncResult a) (AsyncResult b))

Wait for the first of two AsyncSTMs to finish.

waitEither_ :: AsyncSTM a -> AsyncSTM b -> Process ()

Like waitEither, but the result is ignored.

waitBoth :: AsyncSTM a -> AsyncSTM b -> Process (AsyncResult a, AsyncResult b)

Waits for both AsyncSTMs to finish.