Control.Distributed.Process.Platform.ManagedProcess

Description

This module provides a high(er) level API for building complex Process implementations by abstracting out the management of the process' mailbox, reply/response handling, timeouts, process hiberation, error handling and shutdown/stop procedures. It is modelled along similar lines to OTP's gen_server API - http://www.erlang.org/doc/man/gen_server.html.

API Overview

Once started, a generic process will consume messages from its mailbox and pass them on to user defined handlers based on the types received (mapped to those accepted by the handlers) and optionally by also evaluating user supplied predicates to determine which handlers are valid. Each handler returns a ProcessAction which specifies how we should proceed. If none of the handlers is able to process a message (because their types are incompatible) then the process unhandledMessagePolicy will be applied.

The ProcessAction type defines the ways in which a process can respond to its inputs, either by continuing to read incoming messages, setting an optional timeout, sleeping for a while or by stopping. The optional timeout behaves a little differently to the other process actions. If no messages are received within the specified time span, the process timeoutHandler will be called in order to determine the next action.

Generic processes are defined by the ProcessDefinition type, using record syntax. The ProcessDefinition fields contain handlers (or lists of them) for specific tasks. In addtion to the timeoutHandler, a ProcessDefinition may also define a terminateHandler which is called just before the process exits. This handler will be called whenever the process is stopping, i.e., when a callback returns stop as the next action or if an unhandled exit signal or similar asynchronous exception is thrown in (or to) the process itself.

The other handlers are split into two groups: apiHandlers and infoHandlers. The former contains handlers for the cast and call protocols, whilst the latter contains handlers that deal with input messages which are not sent via these API calls (i.e., messages sent using bare send or signals put into the process mailbox by the node controller, such as ProcessMonitorNotification and the like).

The Cast/Call Protocol

Deliberate interactions with the process will usually fall into one of two categories. A cast interaction involves a client sending a message asynchronously and the server handling this input. No reply is sent to the client. On the other hand, a call interaction is a kind of rpc where the client sends a message and waits for a reply.

The expressions given to apiHandlers have to conform to the cast|call protocol. The details of this are, however, hidden from the user. A set of API functions for creating apiHandlers are given instead, which take expressions (i.e., a function or lambda expression) and create the appropriate Dispatcher for handling the cast (or call).

The castcall protocol handlers deal with expected/ inputs. These form the explicit public API for the process, and will usually be exposed by providing module level functions that defer to the cast/call API. For example:

 add :: ProcessId -> Double -> Double -> Double
 add pid x y = call pid (Add x y)

Handling Info Messages

An explicit protocol for communicating with the process can be configured using cast and call, but it is not possible to prevent other kinds of messages from being sent to the process mailbox. When any message arrives for which there are no handlers able to process its content, the UnhandledMessagePolicy will be applied. Sometimes it is desireable to process incoming messages which aren't part of the protocol, rather than let the policy deal with them. This is particularly true when incoming messages are important to the process, but their point of origin is outside the developer's control. Handling signals such as ProcessMonitorNotification is a typical example of this:

 handleInfo_ (\(ProcessMonitorNotification _ _ r) -> say $ show r >> continue_)

Handling Process State

The ProcessDefinition is parameterised by the type of state it maintains. A process that has no state will have the type ProcessDefinition () and can be bootstrapped by evaluating statelessProcess.

All call/cast handlers come in two flavours, those which take the process state as an input and those which do not. Handlers that ignore the process state have to return a function that takes the state and returns the required action. Versions of the various action generating functions ending in an underscore are provided to simplify this:

   statelessProcess {
       apiHandlers = [
         handleCall_   (\(n :: Int) -> return (n * 2))
       , handleCastIf_ (\(c :: String, _ :: Delay) -> c == "timeout")
                       (\("timeout", Delay d) -> timeoutAfter_ d)
       ]
     , timeoutHandler = \_ _ -> stop $ TerminateOther "timeout"
   }

Handling Errors

Error handling appears in several contexts and process definitions can hook into these with relative ease. Only process failures as a result of asynchronous exceptions are supported by the API, which provides several scopes for error handling.

Catching exceptions inside handler functions is no different to ordinary exception handling in monadic code.

   handleCall (\x y ->
                catch (hereBeDragons x y)
                      (\(e :: SmaugTheTerribleException) ->
                           return (Left (show e))))

The caveats mentioned in Control.Distributed.Process.Platform about exit signal handling obviously apply here as well.

Structured Exit Signal Handling

Because Control.Distributed.Process.ProcessExitException is a ubiquitous signalling mechanism in Cloud Haskell, it is treated unlike other asynchronous exceptions. The ProcessDefinition exitHandlers field accepts a list of handlers that, for a specific exit reason, can decide how the process should respond. If none of these handlers matches the type of reason then the process will exit with DiedException why. In addition, a default exit handler is installed for exit signals where the reason == Shutdown, because this is an exit signal used explicitly and extensively throughout the platform. The default behaviour is to gracefully shut down the process, calling the terminateHandler as usual, before stopping with TerminateShutdown given as the final outcome.

Example: How to annoy your supervisor and end up force-killed:

 handleExit  (\state from (sigExit :: Shutdown) -> continue s)

That code is, of course, very silly. Under some circumstances, handling exit signals is perfectly legitimate. Handling of other forms of asynchronous exception is not supported by this API.

If any asynchronous exception goes unhandled, the process will immediately exit without running the terminateHandler. It is very important to note that in Cloud Haskell, link failures generate asynchronous exceptions in the target and these will NOT be caught by the API and will therefore cause the process to exit without running the termination handler callback. If your termination handler is set up to do important work (such as resource cleanup) then you should avoid linking you process and use monitors instead.

Starting server processes

data InitResult s

Return type for and InitHandler expression.

Constructors

InitOk s Delay

forall r . Serializable r => InitFail r
denotes failed initialisation and the reason

type InitHandler a s = a -> Process (InitResult s)

An expression used to initialise a process with its state.

start :: a -> InitHandler a s -> ProcessDefinition s -> Process (Either (InitResult s) TerminateReason)

Starts a gen-process configured with the supplied process definition, using an init handler and its initial arguments. This code will run the Process until completion and return Right TerminateReason *or*, if initialisation fails, return Left InitResult which will be InitFail why.

runProcess :: (ProcessDefinition s -> s -> Delay -> Process TerminateReason) -> a -> InitHandler a s -> ProcessDefinition s -> Process (Either (InitResult s) TerminateReason)

Client interactions

shutdown :: ProcessId -> Process ()

Send a signal instructing the process to terminate. The receive loop which manages the process mailbox will prioritise Shutdown signals higher than any other incoming messages, but the server might be busy (i.e., still in the process of excuting a handler) at the time of sending however, so the caller should not make any assumptions about the timeliness with which the shutdown signal will be handled. If responsiveness is important, a better approach might be to send an exit signal with Shutdown as the reason. An exit signal will interrupt any operation currently underway and force the running process to clean up and terminate.

defaultProcess :: ProcessDefinition s

statelessProcess :: ProcessDefinition ()

A basic, stateless process definition, where the unhandled message policy is set to Terminate, the default timeout handlers does nothing (i.e., the same as calling continue () and the terminate handler is a no-op.

statelessInit :: Delay -> InitHandler () ()

A basic, state unaware InitHandler that can be used with statelessProcess.

call :: forall a b. (Serializable a, Serializable b) => ProcessId -> a -> Process b

Make a synchronous call - will block until a reply is received. The calling process will exit with TerminateReason if the calls fails.

safeCall :: forall a b. (Serializable a, Serializable b) => ProcessId -> a -> Process (Either TerminateReason b)

Safe version of call that returns information about the error if the operation fails. If an error occurs then the explanation will be will be stashed away as (TerminateOther String).

tryCall :: forall s a b. (Addressable s, Serializable a, Serializable b) => s -> a -> Process (Maybe b)

Version of safeCall that returns Nothing if the operation fails. If you need information about *why* a call has failed then you should use safeCall or combine catchExit and call instead.

callAsync :: forall s a b. (Addressable s, Serializable a, Serializable b) => s -> a -> Process (Async b)

Performs a synchronous call to the the given server address, however the call is made out of band and an async handle is returned immediately. This can be passed to functions in the Async API in order to obtain the result.

See Control.Distributed.Process.Platform.Async

callTimeout :: forall s a b. (Addressable s, Serializable a, Serializable b) => s -> a -> TimeInterval -> Process (Maybe b)

Make a synchronous call, but timeout and return Nothing if the reply is not received within the specified time interval.

If the result of the call is a failure (or the call was cancelled) then the calling process will exit, with the AsyncResult given as the reason.

cast :: forall a m. (Addressable a, Serializable m) => a -> m -> Process ()

Sends a cast message to the server identified by ServerId. The server will not send a response. Like Cloud Haskell's send primitive, cast is fully asynchronous and never fails - therefore casting to a non-existent (e.g., dead) server process will not generate an error.

Defining server processes

data ProcessDefinition s

Stores the functions that determine runtime behaviour in response to incoming messages and a policy for responding to unhandled messages.

Constructors

ProcessDefinition

Fields

apiHandlers :: [Dispatcher s]

functions that handle call/cast messages

infoHandlers :: [DeferredDispatcher s]

functions that handle non call/cast messages

exitHandlers :: [ExitSignalDispatcher s]

functions that handle exit signals

timeoutHandler :: TimeoutHandler s

a function that handles timeouts

terminateHandler :: TerminateHandler s

a function that is run just before the process exits

unhandledMessagePolicy :: UnhandledMessagePolicy

how to deal with unhandled messages

type TerminateHandler s = s -> TerminateReason -> Process ()

An expression used to handle process termination.

type TimeoutHandler s = s -> Delay -> Process (ProcessAction s)

An expression used to handle process timeouts.

data ProcessAction s

The action taken by a process after a handler has run and its updated state. See continue timeoutAfter hibernate stop

Constructors

ProcessContinue s
continue with (possibly new) state

ProcessTimeout TimeInterval s
timeout if no messages are received

ProcessHibernate TimeInterval s
hibernate for delay

ProcessStop TerminateReason
stop the process, giving TerminateReason

data ProcessReply s a

Returned from handlers for the synchronous call protocol, encapsulates the reply data and the action to take after sending the reply. A handler can return NoReply if they wish to ignore the call.

type CallHandler a s = s -> a -> Process (ProcessReply s a)

type CastHandler s = s -> Process ()

data UnhandledMessagePolicy

Policy for handling unexpected messages, i.e., messages which are not sent using the call or cast APIs, and which are not handled by any of the handleInfo handlers.

Constructors

Terminate
stop immediately, giving TerminateOther UnhandledInput as the reason

DeadLetter ProcessId
forward the message to the given recipient

Drop
dequeue and then drop/ignore the message

handleCall :: (Serializable a, Serializable b) => (s -> a -> Process (ProcessReply s b)) -> Dispatcher s

handleCallIf

Arguments

:: forall s a b . (Serializable a, Serializable b)

=> Condition s a
predicate that must be satisfied for the handler to run

-> (s -> a -> Process (ProcessReply s b))
a reply yielding function over the process state and input message

-> Dispatcher s

Constructs a call handler from an ordinary function in the Process monad. Given a function f :: (s -> a -> Process (ProcessReply s b)), the expression handleCall f will yield a Dispatcher for inclusion in a Behaviour specification for the GenProcess. Messages are only dispatched to the handler if the supplied condition evaluates to True

handleCallFrom :: forall s a b. (Serializable a, Serializable b) => (s -> Recipient -> a -> Process (ProcessReply s b)) -> Dispatcher s

As handleCall but passes the Recipient to the handler function. This can be useful if you wish to reply later to the caller by, e.g., spawning a process to do some work and have it replyTo caller response out of band. In this case the callback can pass the Recipient to the worker (or stash it away itself) and return noReply.

handleCallFromIf

Arguments

:: forall s a b . (Serializable a, Serializable b)

=> Condition s a
predicate that must be satisfied for the handler to run

-> (s -> Recipient -> a -> Process (ProcessReply s b))
a reply yielding function over the process state, sender and input message

-> Dispatcher s

As handleCallFrom but only runs the handler if the supplied Condition evaluates to True.

handleCast :: Serializable a => (s -> a -> Process (ProcessAction s)) -> Dispatcher s

handleCastIf

Arguments

:: forall s a . Serializable a

=> Condition s a
predicate that must be satisfied for the handler to run

-> (s -> a -> Process (ProcessAction s))
an action yielding function over the process state and input message

-> Dispatcher s

Constructs a cast handler from an ordinary function in the Process monad. Given a function f :: (s -> a -> Process (ProcessAction s)), the expression handleCall f will yield a Dispatcher for inclusion in a Behaviour specification for the GenProcess.

handleInfo :: forall s a. Serializable a => (s -> a -> Process (ProcessAction s)) -> DeferredDispatcher s

Creates a generic input handler (i.e., for recieved messages that are not sent using the cast or call APIs) from an ordinary function in the Process monad.

handleDispatch :: Serializable a => (s -> a -> Process (ProcessAction s)) -> Dispatcher s

Constructs a handler for both call and cast messages. handleDispatch = handleDispatchIf (const True)

handleExit :: forall s a. Serializable a => (s -> ProcessId -> a -> Process (ProcessAction s)) -> ExitSignalDispatcher s

Creates an exit handler scoped to the execution of any and all the registered call, cast and info handlers for the process.

Stateless handlers

action

Arguments

:: forall s a . Serializable a

=> (a -> s -> Process (ProcessAction s))
a function from the input message to a stateless action, cf continue_

-> Dispatcher s

Constructs an action handler. Like handleDispatch this can handle both cast and call messages and you won't know which you're dealing with. This can be useful where certain inputs require a definite action, such as stopping the server, without concern for the state (e.g., when stopping we need only decide to stop, as the terminate handler can deal with state cleanup etc). For example:

action (MyCriticalErrorSignal -> stop_ TerminateNormal)

handleCall_ :: (Serializable a, Serializable b) => (a -> Process b) -> Dispatcher s

Constructs a call handler from a function in the Process monad. The handler expression returns the reply, and the action will be set to continue.

 handleCall_ = handleCallIf_ (const True)

handleCallIf_

Arguments

:: forall s a b . (Serializable a, Serializable b)

=> Condition s a
predicate that must be satisfied for the handler to run

-> (a -> Process b)
a function from an input message to a reply

-> Dispatcher s

Constructs a call handler from an ordinary function in the Process monad. This variant ignores the state argument present in handleCall and handleCallIf and is therefore useful in a stateless server. Messges are only dispatched to the handler if the supplied condition evaluates to True

See handleCall

handleCast_ :: Serializable a => (a -> s -> Process (ProcessAction s)) -> Dispatcher s

Version of handleCast that ignores the server state.

handleCastIf_

Arguments

:: forall s a . Serializable a

=> Condition s a
predicate that must be satisfied for the handler to run

-> (a -> s -> Process (ProcessAction s))
a function from the input message to a stateless action, cf continue_

-> Dispatcher s

Version of handleCastIf that ignores the server state.

Constructing handler results

condition :: forall a b. (Serializable a, Serializable b) => (a -> b -> Bool) -> Condition a b

Creates a Conditon from a function that takes a process state a and an input message b and returns a Bool indicating whether the associated handler should run.

state :: forall s m. Serializable m => (s -> Bool) -> Condition s m

Create a Condition from a function that takes a process state a and returns a Bool indicating whether the associated handler should run.

input :: forall s m. Serializable m => (m -> Bool) -> Condition s m

Creates a Condition from a function that takes an input message m and returns a Bool indicating whether the associated handler should run.

reply :: Serializable r => r -> s -> Process (ProcessReply s r)

Instructs the process to send a reply and continue running.

replyWith :: Serializable m => m -> ProcessAction s -> Process (ProcessReply s m)

Instructs the process to send a reply and evaluate the ProcessAction.

noReply :: Serializable r => ProcessAction s -> Process (ProcessReply s r)

Instructs the process to skip sending a reply and evaluate a ProcessAction

noReply_ :: forall s r. Serializable r => s -> Process (ProcessReply s r)

Continue without giving a reply to the caller - equivalent to continue, but usable in a callback passed to the handleCall family of functions.

haltNoReply_ :: TerminateReason -> Process (ProcessReply s TerminateReason)

Halt process execution during a call handler, without paying any attention to the expected return type.

continue :: s -> Process (ProcessAction s)

Instructs the process to continue running and receiving messages.

continue_ :: s -> Process (ProcessAction s)

Version of continue that can be used in handlers that ignore process state.

timeoutAfter :: TimeInterval -> s -> Process (ProcessAction s)

Instructs the process to wait for incoming messages until TimeInterval is exceeded. If no messages are handled during this period, the timeout handler will be called. Note that this alters the process timeout permanently such that the given TimeInterval will remain in use until changed.

timeoutAfter_ :: TimeInterval -> s -> Process (ProcessAction s)

Version of timeoutAfter that can be used in handlers that ignore process state.

 action (\(TimeoutPlease duration) -> timeoutAfter_ duration)

hibernate :: TimeInterval -> s -> Process (ProcessAction s)

Instructs the process to hibernate for the given TimeInterval. Note that no messages will be removed from the mailbox until after hibernation has ceased. This is equivalent to calling threadDelay.

hibernate_ :: TimeInterval -> s -> Process (ProcessAction s)

Version of hibernate that can be used in handlers that ignore process state.

 action (\(HibernatePlease delay) -> hibernate_ delay)

stop :: TerminateReason -> Process (ProcessAction s)

Instructs the process to terminate, giving the supplied reason. If a valid terminateHandler is installed, it will be called with the TerminateReason returned from this call, along with the process state.

stop_ :: TerminateReason -> s -> Process (ProcessAction s)

Version of stop that can be used in handlers that ignore process state.

 action (\ClientError -> stop_ TerminateNormal)

replyTo :: Serializable m => Recipient -> m -> Process ()