Control.Distributed.Process.Platform.ManagedProcess.Client

Description

Client API

API for client interactions with the process

shutdown :: ProcessId -> Process ()

Send a signal instructing the process to terminate. The receive loop which manages the process mailbox will prioritise Shutdown signals higher than any other incoming messages, but the server might be busy (i.e., still in the process of excuting a handler) at the time of sending however, so the caller should not make any assumptions about the timeliness with which the shutdown signal will be handled. If responsiveness is important, a better approach might be to send an exit signal with Shutdown as the reason. An exit signal will interrupt any operation currently underway and force the running process to clean up and terminate.

call :: forall a b. (Serializable a, Serializable b) => ProcessId -> a -> Process b

Make a synchronous call - will block until a reply is received. The calling process will exit with TerminateReason if the calls fails.

safeCall :: forall a b. (Serializable a, Serializable b) => ProcessId -> a -> Process (Either TerminateReason b)

Safe version of call that returns information about the error if the operation fails. If an error occurs then the explanation will be will be stashed away as (TerminateOther String).

tryCall :: forall s a b. (Addressable s, Serializable a, Serializable b) => s -> a -> Process (Maybe b)

Version of safeCall that returns Nothing if the operation fails. If you need information about *why* a call has failed then you should use safeCall or combine catchExit and call instead.

callAsync :: forall s a b. (Addressable s, Serializable a, Serializable b) => s -> a -> Process (Async b)

Performs a synchronous call to the the given server address, however the call is made out of band and an async handle is returned immediately. This can be passed to functions in the Async API in order to obtain the result.

See Control.Distributed.Process.Platform.Async

callTimeout :: forall s a b. (Addressable s, Serializable a, Serializable b) => s -> a -> TimeInterval -> Process (Maybe b)

Make a synchronous call, but timeout and return Nothing if the reply is not received within the specified time interval.

If the result of the call is a failure (or the call was cancelled) then the calling process will exit, with the AsyncResult given as the reason.

cast :: forall a m. (Addressable a, Serializable m) => a -> m -> Process ()

Sends a cast message to the server identified by ServerId. The server will not send a response. Like Cloud Haskell's send primitive, cast is fully asynchronous and never fails - therefore casting to a non-existent (e.g., dead) server process will not generate an error.