Control.Distributed.Process.Platform.ManagedProcess.Server

Description

Server process API

Server actions

condition :: forall a b. (Serializable a, Serializable b) => (a -> b -> Bool) -> Condition a b

Creates a Conditon from a function that takes a process state a and an input message b and returns a Bool indicating whether the associated handler should run.

state :: forall s m. Serializable m => (s -> Bool) -> Condition s m

Create a Condition from a function that takes a process state a and returns a Bool indicating whether the associated handler should run.

input :: forall s m. Serializable m => (m -> Bool) -> Condition s m

Creates a Condition from a function that takes an input message m and returns a Bool indicating whether the associated handler should run.

reply :: Serializable r => r -> s -> Process (ProcessReply s r)

Instructs the process to send a reply and continue running.

replyWith :: Serializable m => m -> ProcessAction s -> Process (ProcessReply s m)

Instructs the process to send a reply and evaluate the ProcessAction.

noReply :: Serializable r => ProcessAction s -> Process (ProcessReply s r)

Instructs the process to skip sending a reply and evaluate a ProcessAction

noReply_ :: forall s r. Serializable r => s -> Process (ProcessReply s r)

Continue without giving a reply to the caller - equivalent to continue, but usable in a callback passed to the handleCall family of functions.

haltNoReply_ :: TerminateReason -> Process (ProcessReply s TerminateReason)

Halt process execution during a call handler, without paying any attention to the expected return type.

continue :: s -> Process (ProcessAction s)

Instructs the process to continue running and receiving messages.

continue_ :: s -> Process (ProcessAction s)

Version of continue that can be used in handlers that ignore process state.

timeoutAfter :: TimeInterval -> s -> Process (ProcessAction s)

Instructs the process to wait for incoming messages until TimeInterval is exceeded. If no messages are handled during this period, the timeout handler will be called. Note that this alters the process timeout permanently such that the given TimeInterval will remain in use until changed.

timeoutAfter_ :: TimeInterval -> s -> Process (ProcessAction s)

Version of timeoutAfter that can be used in handlers that ignore process state.

 action (\(TimeoutPlease duration) -> timeoutAfter_ duration)

hibernate :: TimeInterval -> s -> Process (ProcessAction s)

Instructs the process to hibernate for the given TimeInterval. Note that no messages will be removed from the mailbox until after hibernation has ceased. This is equivalent to calling threadDelay.

hibernate_ :: TimeInterval -> s -> Process (ProcessAction s)

Version of hibernate that can be used in handlers that ignore process state.

 action (\(HibernatePlease delay) -> hibernate_ delay)

stop :: TerminateReason -> Process (ProcessAction s)

Instructs the process to terminate, giving the supplied reason. If a valid terminateHandler is installed, it will be called with the TerminateReason returned from this call, along with the process state.

stop_ :: TerminateReason -> s -> Process (ProcessAction s)

Version of stop that can be used in handlers that ignore process state.

 action (\ClientError -> stop_ TerminateNormal)

replyTo :: Serializable m => Recipient -> m -> Process ()

Server handler/callback creation

handleCall :: (Serializable a, Serializable b) => (s -> a -> Process (ProcessReply s b)) -> Dispatcher s

handleCallIf

Arguments

:: forall s a b . (Serializable a, Serializable b)

=> Condition s a
predicate that must be satisfied for the handler to run

-> (s -> a -> Process (ProcessReply s b))
a reply yielding function over the process state and input message

-> Dispatcher s

Constructs a call handler from an ordinary function in the Process monad. Given a function f :: (s -> a -> Process (ProcessReply s b)), the expression handleCall f will yield a Dispatcher for inclusion in a Behaviour specification for the GenProcess. Messages are only dispatched to the handler if the supplied condition evaluates to True

handleCallFrom :: forall s a b. (Serializable a, Serializable b) => (s -> Recipient -> a -> Process (ProcessReply s b)) -> Dispatcher s

As handleCall but passes the Recipient to the handler function. This can be useful if you wish to reply later to the caller by, e.g., spawning a process to do some work and have it replyTo caller response out of band. In this case the callback can pass the Recipient to the worker (or stash it away itself) and return noReply.

handleCallFromIf

Arguments

:: forall s a b . (Serializable a, Serializable b)

=> Condition s a
predicate that must be satisfied for the handler to run

-> (s -> Recipient -> a -> Process (ProcessReply s b))
a reply yielding function over the process state, sender and input message

-> Dispatcher s

As handleCallFrom but only runs the handler if the supplied Condition evaluates to True.

handleCast :: Serializable a => (s -> a -> Process (ProcessAction s)) -> Dispatcher s

handleCastIf

Arguments

:: forall s a . Serializable a

=> Condition s a
predicate that must be satisfied for the handler to run

-> (s -> a -> Process (ProcessAction s))
an action yielding function over the process state and input message

-> Dispatcher s

Constructs a cast handler from an ordinary function in the Process monad. Given a function f :: (s -> a -> Process (ProcessAction s)), the expression handleCall f will yield a Dispatcher for inclusion in a Behaviour specification for the GenProcess.

handleInfo :: forall s a. Serializable a => (s -> a -> Process (ProcessAction s)) -> DeferredDispatcher s

Creates a generic input handler (i.e., for recieved messages that are not sent using the cast or call APIs) from an ordinary function in the Process monad.

handleDispatch :: Serializable a => (s -> a -> Process (ProcessAction s)) -> Dispatcher s

Constructs a handler for both call and cast messages. handleDispatch = handleDispatchIf (const True)

handleExit :: forall s a. Serializable a => (s -> ProcessId -> a -> Process (ProcessAction s)) -> ExitSignalDispatcher s

Creates an exit handler scoped to the execution of any and all the registered call, cast and info handlers for the process.

Stateless handlers

action

Arguments

:: forall s a . Serializable a

=> (a -> s -> Process (ProcessAction s))
a function from the input message to a stateless action, cf continue_

-> Dispatcher s

Constructs an action handler. Like handleDispatch this can handle both cast and call messages and you won't know which you're dealing with. This can be useful where certain inputs require a definite action, such as stopping the server, without concern for the state (e.g., when stopping we need only decide to stop, as the terminate handler can deal with state cleanup etc). For example:

action (MyCriticalErrorSignal -> stop_ TerminateNormal)

handleCall_ :: (Serializable a, Serializable b) => (a -> Process b) -> Dispatcher s

Constructs a call handler from a function in the Process monad. The handler expression returns the reply, and the action will be set to continue.

 handleCall_ = handleCallIf_ (const True)

handleCallIf_

Arguments

:: forall s a b . (Serializable a, Serializable b)

=> Condition s a
predicate that must be satisfied for the handler to run

-> (a -> Process b)
a function from an input message to a reply

-> Dispatcher s

Constructs a call handler from an ordinary function in the Process monad. This variant ignores the state argument present in handleCall and handleCallIf and is therefore useful in a stateless server. Messges are only dispatched to the handler if the supplied condition evaluates to True

See handleCall

handleCast_ :: Serializable a => (a -> s -> Process (ProcessAction s)) -> Dispatcher s

Version of handleCast that ignores the server state.

handleCastIf_

Arguments

:: forall s a . Serializable a

=> Condition s a
predicate that must be satisfied for the handler to run

-> (a -> s -> Process (ProcessAction s))
a function from the input message to a stateless action, cf continue_

-> Dispatcher s

Version of handleCastIf that ignores the server state.