Control.Distributed.Process.Platform.Test

Description

This module provides basic building blocks for testing Cloud Haskell programs.

Documentation

type TestResult a = MVar a

A mutable cell containing a test result.

noop :: Process ()

Does exactly what it says on the tin, doing so in the Process monad.

stash :: TestResult a -> a -> Process ()

Stashes a value in our TestResult using putMVar

data Ping

A simple Ping signal

Constructors

Ping

ping :: ProcessId -> Process ()

data TestProcessControl

Control signals used to manage test processes

startTestProcess :: Process () -> Process ProcessId

Starts a test process on the local node.

runTestProcess :: Process () -> Process ()

Runs a test process around the supplied proc, which is executed whenever the outer process loop receives a Go signal.

testProcessGo :: ProcessId -> Process ()

Tell a test process to continue executing

testProcessStop :: ProcessId -> Process ()

Tell a test process to stop (i.e., terminate)

testProcessReport :: ProcessId -> Process ()

Tell a test process to send a report (message) back to the calling process

tryRunProcess :: LocalNode -> Process () -> IO ()

tryForkProcess :: LocalNode -> Process () -> IO ProcessId