Control.Distributed.Process.Platform.Time

Description

This module provides facilities for working with time delays and timeouts. The type Timeout and the timeout family of functions provide mechanisms for working with threadDelay-like behaviour operates on microsecond values.

The TimeInterval and TimeUnit related functions provide an abstraction for working with various time intervals and the Delay type provides a corrolary to timeout that works with these.

Documentation

microSeconds :: Int -> TimeInterval

milliSeconds :: Int -> TimeInterval

given a number, produces a TimeInterval of milliseconds

seconds :: Int -> TimeInterval

given a number, produces a TimeInterval of seconds

minutes :: Int -> TimeInterval

given a number, produces a TimeInterval of minutes

hours :: Int -> TimeInterval

given a number, produces a TimeInterval of hours

asTimeout :: TimeInterval -> Int

converts the supplied TimeInterval to milliseconds

after :: Int -> TimeUnit -> Int

Convenience for making timeouts; e.g.,

 receiveTimeout (after 3 Seconds) [ match (\"ok" -> return ()) ]

within :: Int -> TimeUnit -> TimeInterval

Convenience for making TimeInterval; e.g.,

 let ti = within 5 Seconds in .....

timeToMs :: TimeUnit -> Int -> Int

converts the supplied TimeUnit to microseconds

data TimeInterval

data TimeUnit

Defines the time unit for a Timeout value

Constructors

Days

Hours

Minutes

Seconds

Millis

Micros

data Delay

Constructors

Delay TimeInterval

Infinity

type Timeout = Maybe Int

Represents a timeout in terms of microseconds, where Nothing stands for infinity and Just 0, no-delay.

data TimeoutNotification

Send to a process when a timeout expires.

Constructors

TimeoutNotification Tag

timeout :: Int -> Tag -> ProcessId -> Process ()

Sends the calling process TimeoutNotification tag after time microseconds

infiniteWait :: Timeout

Constructs an inifinite Timeout.

noWait :: Timeout

Constructs a no-wait Timeout