Control.Distributed.Process.Platform.Timer

Description

Provides an API for running code or sending messages, either after some initial delay or periodically, and for cancelling, re-setting and/or flushing pending timers.

Documentation

type TimerRef = ProcessId

an opaque reference to a timer

data Tick

represents a tick event that timers can generate

Constructors

Tick

sleep :: TimeInterval -> Process ()

blocks the calling Process for the specified TimeInterval. Note that this function assumes that a blocking receive is the most efficient approach to acheiving this, however the runtime semantics (particularly with regards scheduling) should not differ from threadDelay in practise.

sendAfter :: Serializable a => TimeInterval -> ProcessId -> a -> Process TimerRef

starts a timer which sends the supplied message to the destination process after the specified time interval.

runAfter :: TimeInterval -> Process () -> Process TimerRef

runs the supplied process action(s) after t has elapsed

exitAfter :: Serializable a => TimeInterval -> ProcessId -> a -> Process TimerRef

calls exit pid reason after t has elapsed

killAfter :: TimeInterval -> ProcessId -> String -> Process TimerRef

kills the specified process after t has elapsed

startTimer :: Serializable a => TimeInterval -> ProcessId -> a -> Process TimerRef

starts a timer that repeatedly sends the supplied message to the destination process each time the specified time interval elapses. To stop messages from being sent in future, cancelTimer can be called.

ticker :: TimeInterval -> ProcessId -> Process TimerRef

sets up a timer that sends Tick repeatedly at intervals of t

periodically :: TimeInterval -> Process () -> Process TimerRef

runs the supplied process action(s) repeatedly at intervals of t

resetTimer :: TimerRef -> Process ()

resets a running timer. Note: Cancelling a timer does not guarantee that all its messages are prevented from being delivered to the target process. Also note that resetting an ongoing timer (started using the startTimer or periodically functions) will only cause the current elapsed period to time out, after which the timer will continue running. To stop a long-running timer permanently, you should use cancelTimer instead.

cancelTimer :: TimerRef -> Process ()

permanently cancels a timer

flushTimer :: (Serializable a, Eq a) => TimerRef -> a -> Delay -> Process ()

cancels a running timer and flushes any viable timer messages from the process' message queue. This function should only be called by the process expecting to receive the timer's messages!